Time: 0 ms (100%), Space: 17.9 MB (23.95%) - LeetHub

+++ 0152-maximum-product-subarray/0152-maximum-product-subarray.cpp
@@ -2,12 +2,12 @@ class Solution {
 public:
     int maxProduct(vector<int>& nums) {
         if(nums.size()==1) return nums[0];
-        priority_queue<long long> pq;
-        long long prod=1,res=nums[0],zc=0;;
+        priority_queue<int> pq;
+        int prod=1,res=nums[0],zc=0;;
         for(auto& i:nums){
             prod*=i;
             if(prod<0){
-                long long temp = prod;
+                int temp = prod;
                 if(pq.size()) temp/=pq.top();
                 res=max({res,temp,prod});
                 pq.push(prod);
@@ -15,7 +15,7 @@ public:
             else if(!prod){
                 res=max(res,prod);
                 prod=1;
-                pq=priority_queue<long long>();
+                pq=priority_queue<int>();
             }
             else res=max(res,prod);
         }
